Change package directory
