--- src/main/java/Launcher.java
@@ -1,8 +1,0 @@
-import javafx.application.Application;
-import util.windows.Main;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
