+++ src/main/java/duke/Launcher.java
@@ -1,0 +1,10 @@
+package duke;
+
+import javafx.application.Application;
+import duke.util.windows.Main;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
