+++ src/main/java/duke/util/commands/ExitCommand.java
@@ -1,0 +1,32 @@
+package duke.util.commands;
+
+
+/**
+ * When this command runs, the Duke stops working.
+ *
+ */
+public class ExitCommand implements Command {
+
+
+    private static boolean isClosed = false;
+
+    /**
+     * The constructor of the ExitCommand.
+     *
+     */
+    public ExitCommand() {
+    }
+
+    @Override
+    public String execute() {
+
+        this.isClosed = true;
+        return "";
+    }
+
+    public static boolean isExit() {
+        return ExitCommand.isClosed;
+    }
+
+
+}
