+++ src/main/java/duke/util/tasks/DukeException.java
@@ -1,0 +1,19 @@
+package duke.util.tasks;
+
+/**
+ * DukeException, an exception that arose from inaccurate input.
+ *
+ */
+public class DukeException extends Exception {
+
+    public DukeException() {
+        super("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
+    }
+
+
+    public DukeException(String message) {
+        super(message);
+    }
+
+
+}
