--- src/main/java/util/commands/Command.java
@@ -1,15 +1,0 @@
-package util.commands;
-
-
-import util.tasks.DukeException;
-
-
-@FunctionalInterface
-public interface Command {
-    /**
-     * The execution of a command.
-     *
-     * @throws DukeException When the command activity throws the exception.
-     */
-    public String execute() throws DukeException;
-}
