--- src/main/java/util/tasks/ToDo.java
@@ -1,38 +1,0 @@
-package util.tasks;
-
-public class ToDo extends Task {
-
-
-    private static String label = "[T]";
-
-
-    private ToDo(String s) {
-        super(s.trim());
-
-    }
-
-    /**
-     * The factory method of the todo class
-     *
-     * @param s The name of the todo task.
-     * @return The Task with the corresponding name.
-     */
-    public static Task of(String s) {
-        return new ToDo(s);
-    }
-
-    @Override
-    public String toString() {
-        return this.label + super.toString();
-    }
-
-    @Override
-    public String encode() {
-        //String indicating whether this task is done or not.
-        String d = this.isDone()
-                ? Task.DONE
-                : Task.NOTDONE;
-
-        return Task.Label.T + Task.DELIMITER + d + Task.DELIMITER + this.name;
-    }
-}
