--- src/main/java/util/windows/Main.java
@@ -1,37 +1,0 @@
-package util.windows;
-
-import java.io.IOException;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-import util.controller.Duke;
-
-/**
- * A GUI for Duke using FXML.
- */
-public class Main extends Application {
-
-    private static final String saveFilePath = "save.txt";
-    private static final String tempFilePath = "temp.txt";
-    private Duke duke = new Duke(saveFilePath, tempFilePath);
-
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            stage.setMinHeight(590);
-            stage.setMinWidth(770);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            TaskInputWindow.setDuke(duke);
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-}
